`diff`: Fix name -> index conversion
This fixes the conversion from column name -> index for `diff` options
`--key` and `--sort-columns`, which is part of #2443.

The issue was that `enumeration idx` and `1` had been added to the already
correct result (obtained via `.position` iterator method).
So now, only `.position` is used to get the correct idx of a column name.

Before this change, one test has tried to validate the logic of providing
column names instead of indices, but the test itself was wrong, as the
diff result that was validated did not have the correct sort order.
This test is now also fixed.

Additionally, this provides some more tests for error conditions
regarding name -> index conversion.